+++ test/integration/admin/invoices/details_test.rb
@@ -9,11 +9,18 @@ class AdminInvoiceDetailsTest < ActionDispatch::IntegrationTest
   end
 
   def test_downloads_invoice
-    get download_admin_invoice_path(@invoice)
+    pdf_mock = Minitest::Mock.new
+    pdf_mock.expect(:to_pdf, "%PDF-1.4 stub")
 
-    assert_response :ok
-    assert_equal 'application/pdf', response.headers['content-type']
-    assert_equal %(attachment; filename="#{@invoice.filename}"; filename*=UTF-8''#{@invoice.filename}), response.headers['content-disposition']
-    assert_not_empty response.body
+    PDFKit.stub(:new, pdf_mock) do
+      get download_admin_invoice_path(@invoice)
+
+      assert_response :ok
+      assert_equal 'application/pdf', response.headers['content-type']
+      assert_equal %(attachment; filename="#{@invoice.filename}"; filename*=UTF-8''#{@invoice.filename}), response.headers['content-disposition']
+      assert_not_empty response.body
+    end
+
+    pdf_mock.verify
   end
 end